refactor select framework page
